separate zone-list
