+++ catalystwan/endpoints/configuration/policy/list/zone.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import ZoneList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.zone import ZoneList, ZoneListEditPayload, ZoneListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class ZoneListEditPayload(ZoneList, PolicyListId):
-    pass
-
-
-class ZoneListInfo(ZoneList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyZoneList(APIEndpoints, PolicyListEndpoints):
