+++ catalystwan/models/policy/lists.py
@@ -1,11 +1,8 @@
 # Copyright 2023 Cisco Systems, Inc. and its affiliates
 
-from typing import Any, List, Literal, Optional, Set
+from typing import Any, List, Optional
 
 from pydantic import BaseModel, Field
-
-from catalystwan.models.common import InterfaceType
-from catalystwan.models.policy.lists_entries import ZoneListEntry
 
 
 class PolicyListBase(BaseModel):
@@ -22,14 +19,3 @@ class PolicyListBase(BaseModel):
             del self.entries[1:]
         else:
             self.entries.append(entry)
-
-
-class ZoneList(PolicyListBase):
-    type: Literal["zone"] = "zone"
-    entries: List[ZoneListEntry] = []
-
-    def assign_vpns(self, vpns: Set[int]) -> None:
-        self.entries = [ZoneListEntry(vpn=(vpn, None)) for vpn in vpns]
-
-    def assign_interfaces(self, ifs: Set[InterfaceType]) -> None:
-        self.entries = [ZoneListEntry(interface=interface) for interface in ifs]
